chore: move some code to the lib so we can share it between tests (sigh)
